+++ R/preprocess.R
@@ -863,21 +863,6 @@ run_preprocess_multPeriods <- function(yname,
       }
     }
 
-    # check if bootstrap is on
-    if (!boot){
-      warning("Clustered SEs are only available when boot=TRUE. Setting boot=TRUE and cband=TRUE for bootstrapped standard errors.")
-      boot <- TRUE
-      args$boot <- boot
-      cband <- TRUE
-      args$cband <- cband
-
-      # adding boot reps too
-      if (is.null(nboot)){
-        warning("Number of bootstrap samples not specified. Defaulting to 999 reps.")
-        nboot <- 999
-        args$nboot <- nboot
-      }
-    }
   }
 
   # Flag for parallel and cores